Cleaning up and moving files into related folders
