+++ test/root_component_test.dart
@@ -4,8 +4,8 @@ import 'dart:async';
 import 'package:angular/angular.dart';
 import 'package:angular_test/angular_test.dart';
 import 'package:dv360_excel_plugin/root_component.dart';
-import 'package:dv360_excel_plugin/src/excel.dart';
-import 'package:dv360_excel_plugin/src/gapi.dart';
+import 'package:dv360_excel_plugin/src/javascript_api/excel_api.dart';
+import 'package:dv360_excel_plugin/src/javascript_api/google_api.dart';
 import 'package:intl/intl.dart';
 import 'package:mockito/mockito.dart';
 import 'package:pageloader/html.dart';
